% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hapmap.R
\name{convert.hapmap}
\alias{convert.hapmap}
\title{Change the encoding of a hapmap file}
\usage{
convert.hapmap(hapmap, encoding = "rrBLUP")
}
\arguments{
\item{hapmap}{A TASSEL-encoded or rrBLUP-encoded hapmap file read in to R as
a data.frame. See \code{Details} for information on file format.}

\item{encoding}{The desired output encoding. Either \code{"rrBLUP"} or
\code{"TASSEL"}. See \code{Details} for information on file format}
}
\value{
A \code{data.frame} of a hapmap encoded in the designated format.
}
\description{
Changes the encoding of a TASSEL hapmap to a rrBLUP hapmap and vice-versa
}
\details{
The TASSEL format is as such:
The first row is column names. The first 4 columns are marker name, alleles,
chromosome, and position, respectively. The next 7 column are additional information for
TASSEL. The remaining columns are samples. Genotypes are encoded in
diploid format (i.e. AA, AC, CC) with "NN" denoting missing data.

The rrBLUP format is as such:
The first row is column names. The first 4 columns are marker name, alleles,
chromosome, and position, respectively. The next 7 column are additional information for
TASSEL. The remaining columns are samples. Genotypes are encoded in {1, 0, -1}
format where 1 is homozygous for the first allele, 0 is heterozygous, and -1 is
homozygous for the second allele. Missing data is denoted with \code{NA}.
}

