% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{convert_hapmap}
\alias{convert_hapmap}
\title{Convert a hapmap to a genotype matrix}
\usage{
convert_hapmap(hapmap)
}
\arguments{
\item{hapmap}{A hapmap data.frame using the rrBLUP or TASSEL encoding.}
}
\description{
Convert a hapmap to a genotype matrix
}
