% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_mat.R
\name{G_mat}
\alias{G_mat}
\title{Calculate a genomic relationship matrix from marker information}
\usage{
G_mat(
  X,
  base.pop = NULL,
  min.MAF = NULL,
  max.missing = NULL,
  impute.method = c("mean", "EM", "pass"),
  tol = 0.02,
  shrink = FALSE
)
}
\arguments{
\item{X}{A matrix (\eqn{n \times m}) of marker data on \eqn{n} entries and
\eqn{m} biallelic markers, coded as {-1, 0, 1}. Fractional (imputed)
and missing values (NA) are allowed. Must have row (entry) names and column
(marker) names.}

\item{base.pop}{A \code{character} vector of entry names that make up the
base population (i.e. unselected) for calculating marker allele frequencies
(see \emph{Details}).}

\item{min.MAF}{See \code{\link[rrBLUP]{A.mat}}}

\item{max.missing}{See \code{\link[rrBLUP]{A.mat}}}

\item{impute.method}{The method for imputation. Can be "mean", "EM", or "pass".
See \code{\link{rrBLUP}[A.mat]} for "mean" or "EM". If "pass", markers
are assumed imputed.}

\item{tol}{See \code{\link[rrBLUP]{A.mat}}}

\item{shrink}{See \code{\link[rrBLUP]{A.mat}}}
}
\description{
Calculates the genomic relationship matrix \code{G} using methods described
by Van Raden (2008).
}
