% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_PCA.R
\name{G_PCA}
\alias{G_PCA}
\title{Plot the principal components of a genomic relationship matrix}
\usage{
G_PCA(G, x.PC, y.PC, col = NULL, scree = FALSE, return.data = FALSE)
}
\arguments{
\item{G}{A genomic relationship \code{matrix} of dimensions n x n, where n
is the number of entries.}

\item{x.PC}{The principal component to plot on the x-axis.}

\item{y.PC}{The principal component to plot on the y-axis.}

\item{col}{A \code{factor} of categories by which to color the plot of
principal componts. Must be of length n.}

\item{scree}{A \code{logical} indicating whether to display a scree plot.}

\item{return.data}{A \code{logical} indicating whether to return the results
of the singular value decomposition.}
}
\description{
This function takes a genomic relationship matrix (i.e. one generated by
the \code{A.mat} function in \code{rrBLUP}), performs a singular value
decomposition, and plots the designated principal components.
}
