% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_summary.R
\name{geno_summary}
\alias{geno_summary}
\title{Summarize a marker matrix}
\usage{
geno_summary(genos.mat, print.plot = TRUE)
}
\arguments{
\item{genos.mat}{A n x m marker matrix where n is the number of entries and m
is the number of markers.}

\item{print.plot}{Logical. Should plots of the statistics be displayed?}
}
\description{
Summarize a marker matrix
}
