% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_predict_quick.R
\name{pop_predict_quick}
\alias{pop_predict_quick}
\title{Quicker procedures for PopVar}
\usage{
pop_predict_quick(G.in, y.in, map.in, crossing.table, parents, tail.p,
  model = c("RRBLUP"), map.function = c("haldane", "kosambi", "cf",
  "morgan"))
}
\arguments{
\item{G.in}{TBD}

\item{y.in}{TBD}

\item{map.in}{TBD}

\item{crossing.table}{TBD}

\item{parents}{TBD}

\item{tail.p}{TBD}

\item{model}{TBD}

\item{map.function}{TBD}
}
\description{
Calculates the expected mean, genetic variance, and superior progeny mean of a bi-parental population
based on genomewide marker effects.
}
\examples{



}
