% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ranef_model_matrix}
\alias{ranef_model_matrix}
\title{Create a random effect incidence matrix}
\usage{
ranef_model_matrix(random, data, vcov, sparse = TRUE)
}
\arguments{
\item{random}{A \code{formula} specifying the the random effect terms of the
linear mixed-model. Must be preceded by a \code{~} and can include the modifier
verbs \code{\link{at}[sommer]}, \code{\link{us}[sommer]}, and \code{\link{diag}[base]}
to specify different structures. If a random effect term has an accompanying
variance-covariance matrix (see \code{vcov}), the term should be nested using
the function \code{\link{g}[sommer]}}

\item{data}{A \code{data.frame} that includes the response and random effect
predictors. If not supplied, the function will look in the current running
environment.}

\item{vcov}{A named list of variance-covariance matrices that correspond to
the random effect terms that are specified with the \code{\link[g]{sommer}}
function. If not passed, an identity matrix is supplied.}
}
\value{
A nested list of model matrices and variance-covariance matrices. For each
random effect term, the incidence matrix \code{Z} is returned along with its
accompanying variance-covariance matrix \code{K}.
}
\description{
Creates an incidence matrix for the random effects of a mixed-model
}
