% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{resid_model_matrix}
\alias{resid_model_matrix}
\title{Create a residual effect covariance matrix}
\usage{
resid_model_matrix(resid, data, sparse = TRUE)
}
\arguments{
\item{resid}{A \code{formula} specifying the the structure of the residual
effects in the linear mixed-model. Must be preceded by a \code{~} and can
include the modifier verbs \code{\link[at]{sommer}}, \code{\link[us]{sommer}},
and \code{\link[diag]{base}} to specify different structures. The formulae
can take the same form as those in \code{\link[ranef_model_matrix]{gws}},
except the term must be "units".}

\item{data}{A \code{data.frame} that includes the response and predictors.
If not supplied, the function will look in the current running environment.}
}
\value{
A nested list of model matrices.
}
\description{
Creates an covariance matrix for the random effects of a mixed-model
}
