% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2tp_bopa_unfilt.R
\docType{data}
\name{s2tp_bopa_unfilt}
\alias{s2tp_bopa_unfilt}
\title{Sample unfiltered hapmap}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7369 rows and 187 columns.}
\source{
https://triticeaetoolbox.org/barley/
}
\usage{
s2tp_bopa_unfilt
}
\description{
An unfiltered hapmap file of 7369 SNP marker genotypes for 183 inbred two-row
spring barley lines
}
\keyword{datasets}
